Extract common MDL compilation utilities
Centralize execution-context diagnostics, failure handling, and target
argument-block capture in MdlUtils. Use the same implementations from
material compilation, BSDF compilation, and the MDL SDK tests.
