+++ examples/DemandLoading/DemandPbrtScene/ProgramGroups.cpp
@@ -15,6 +15,7 @@
 #include "DemandPbrtScene/MdlHandleTypes.h"
 #include "DemandPbrtScene/MdlSdkSession.h"
 #include "DemandPbrtScene/MdlShaderCache.h"
+#include "DemandPbrtScene/MdlUtils.h"
 #endif
 #include "DemandPbrtScene/Options.h"
 #include "DemandPbrtScene/Params.h"
@@ -66,7 +67,6 @@
 #include <mutex>
 #include <set>
 #endif
-#include <sstream>
 #include <stdexcept>
 #include <string>
 #ifdef OTK_USE_MDL
@@ -217,85 +217,6 @@ class MdlOptixModuleCache
     std::mutex                         m_mutex;
     std::map<std::string, OptixModule> m_modules;
 };
-
-std::string describeContextMessages( const mi::neuraylib::IMdl_execution_context* context )
-{
-    if( !context )
-        return {};
-
-    std::ostringstream out;
-    for( mi::Size i = 0; i < context->get_messages_count(); ++i )
-    {
-        MessageHandle message( context->get_message( i ) );
-        if( message.is_valid_interface() )
-            out << message->get_string() << '\n';
-    }
-    return out.str();
-}
-
-[[noreturn]] void failMdl( const std::string& message, const mi::neuraylib::IMdl_execution_context* context = nullptr )
-{
-    const std::string contextMessages{ describeContextMessages( context ) };
-    throw std::runtime_error( contextMessages.empty() ? message : message + ":\n" + contextMessages );
-}
-
-void requireMdl( bool condition, const std::string& message, const mi::neuraylib::IMdl_execution_context* context = nullptr )
-{
-    if( !condition )
-    {
-        failMdl( message, context );
-    }
-}
-
-MdlTargetArgumentBlock captureMdlTargetArgumentBlock( const mi::neuraylib::ITarget_code*       targetCode,
-                                                      const mi::neuraylib::ICompiled_material* compiledMaterial,
-                                                      mi::neuraylib::IMdl_execution_context*   context )
-{
-    requireMdl( targetCode != nullptr, "Cannot capture MDL argument block without target code", context );
-    requireMdl( compiledMaterial != nullptr, "Cannot capture MDL argument block without a compiled material", context );
-    requireMdl( targetCode->get_callable_function_count() > 0U,
-                "Cannot capture MDL argument block without callable functions", context );
-
-    const mi::Size argumentBlockIndex{ targetCode->get_callable_function_argument_block_index( 0U ) };
-    if( argumentBlockIndex == ~mi::Size( 0 ) )
-    {
-        return MdlTargetArgumentBlock{};
-    }
-
-    TargetArgumentBlockHandle argumentBlock( targetCode->get_argument_block( argumentBlockIndex ) );
-    requireMdl( argumentBlock.is_valid_interface(), "MDL target code did not expose an argument block", context );
-
-    TargetValueLayoutHandle layout( targetCode->get_argument_block_layout( argumentBlockIndex ) );
-    requireMdl( layout.is_valid_interface(), "MDL target code did not expose an argument block layout", context );
-
-    MdlTargetArgumentBlock result;
-    result.data.assign( argumentBlock->get_data(), argumentBlock->get_data() + argumentBlock->get_size() );
-
-    const mi::Size parameterCount{ compiledMaterial->get_parameter_count() };
-    requireMdl( layout->get_num_elements() >= parameterCount,
-                "MDL argument block layout has fewer entries than the compiled material", context );
-    for( mi::Size i = 0; i < parameterCount; ++i )
-    {
-        const char* const name = compiledMaterial->get_parameter_name( i );
-        requireMdl( name != nullptr, "MDL compiled material exposed a null parameter name", context );
-
-        const mi::neuraylib::Target_value_layout_state state{ layout->get_nested_state( i ) };
-        requireMdl( state.m_state_offs != ~mi::Uint32( 0 ),
-                    "MDL argument block layout did not expose parameter state for " + std::string{ name }, context );
-
-        mi::neuraylib::IValue::Kind kind{};
-        mi::Size                    size{};
-        const mi::Size              offset{ layout->get_layout( kind, size, state ) };
-        requireMdl( offset != ~mi::Size( 0 ),
-                    "MDL argument block layout did not expose parameter offset for " + std::string{ name }, context );
-        requireMdl( offset + size <= result.data.size(),
-                    "MDL argument block layout parameter exceeds block size for " + std::string{ name }, context );
-
-        result.parameters.push_back( MdlTargetArgumentBlockParameter{
-            name, static_cast<unsigned int>( kind ), static_cast<std::size_t>( offset ), static_cast<std::size_t>( size ) } );
-    }
-    return result;
-}
 
 bool isPtxIdentifierChar( char value )
 {
